finbot/backend/core/config.py
import os
from dotenv import load_dotenv
from langchain_huggingface import HuggingFaceEmbeddings

load_dotenv()

# ENV values
SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = os.getenv("ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 30))
OPENROUTER_API_KEY = os.getenv("OPENROUTER_API_KEY")

# Paths
CHROMA_DIR = "chroma_db"
DATA_DIR = "data"

# Shared embedding model instance
embedding_model = HuggingFaceEmbeddings(model_name="sentence-transformers/all-MiniLM-L6-v2")

