finbot/requirements.txt
fastapi
streamlit
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
langchain-huggingface
chromadb==0.4.22
sentence-transformers
pandas
passlib[bcrypt]
python-multipart
requests
PyJWT
numpy==1.26.4
python-jose[cryptography]
bcrypt==4.0.1
